+++ changelog.md
@@ -1,5 +1,41 @@
 # Additional Block Styles Changelog
 
+## 1.1.0, 20220509
+
+### Added
+- Plugin options
+
+### Updated
+- Media & Text image padding and alignment
+- Localization
+
+### Fixed
+- Style bugs with latest Gutenberg plugin (single column alignment and separator background color)
+- Style specificity
+
+### File updates
+	additional-block-styles.php
+	changelog.md
+	readme.txt
+	assets/scss/global.scss
+	assets/scss/_tools/_function-round.scss
+	assets/scss/styles/alignleft.scss
+	assets/scss/styles/alignright.scss
+	assets/scss/styles/double-line.scss
+	assets/scss/styles/gradient.scss
+	assets/scss/styles/hidden-on-tablet.scss
+	assets/scss/styles/media-on-top.scss
+	assets/scss/styles/stacked-on-tablet.scss
+	assets/scss/styles/zigzag.scss
+	assets/scss/styles/includes/_media-on-top.scss
+	includes/Assets.php
+	includes/Autoload.php
+	includes/Loader.php
+	includes/Options.php
+	includes/parts/admin-page.php
+	languages/*.*
+
+
 ## 1.0.0, 20220330
 
 - Initial release.